Added CreatePostTest and changed the the changed ServerManagerTest to ConnectionTest
